back ported about dialog
