+++ app/controllers/navigation.js
@@ -1,11 +1,14 @@
 import Ember from "ember";
+import HospitalRunVersion from 'hospitalrun/mixins/hospitalrun-version';
+import ModalHelper from 'hospitalrun/mixins/modal-helper';
 import ProgressDialog from "hospitalrun/mixins/progress-dialog";
 import UserSession from "hospitalrun/mixins/user-session";
 import Navigation from "hospitalrun/mixins/navigation";
-export default Ember.Controller.extend(ProgressDialog, UserSession, Navigation, {
-    
+export default Ember.Controller.extend(HospitalRunVersion, ModalHelper, ProgressDialog, UserSession, Navigation, {
+
     needs: 'application',
     allowSearch: false,
+    configs: Ember.computed.alias('controllers.application.model'),
     currentSearchText: null,
     currentRouteName: Ember.computed.alias('controllers.application.currentRouteName'),
     progressTitle: 'Searching',
@@ -14,6 +17,16 @@ export default Ember.Controller.extend(ProgressDialog, UserSession, Navigation, 
     currentOpenNav: null,
 
     actions: {
+        about: function() {
+            var configs = this.get('configs'),
+                version = this.get('version'),
+                message = `Version: ${version}`,
+                siteInfo = configs.findBy('id','site_information');
+            if (!Ember.isEmpty(siteInfo)) {
+                message += ` Site Info: ${siteInfo.get('value')}`;
+            }
+            this.displayAlert('About HospitalRun', message);
+        },
         search: function() {
             if (this.allowSearch && this.searchRoute) {
                 var currentRouteName = this.get('currentRouteName'),
@@ -40,7 +53,7 @@ export default Ember.Controller.extend(ProgressDialog, UserSession, Navigation, 
         toggleSettings: function() {
             this.toggleProperty('isShowingSettings');
         },
-        
+
         closeSettings: function() {
             this.set('isShowingSettings', false);
         }
