+++ app/mixins/hospitalrun-version.js
@@ -1,0 +1,4 @@
+import Ember from 'ember';
+export default Ember.Mixin.create({
+    version: '0.8.7.1'
+});
