+++ app/routes/application.js
@@ -9,13 +9,38 @@ var ApplicationRoute = Ember.Route.extend(Ember.SimpleAuth.ApplicationRouteMixin
             } else {
                 this._super();
             }
+        },
+        closeModal: function() {
+            this.disconnectOutlet({
+                parentView: 'application',
+                outlet: 'modal'
+            });
+        },
+        /**
+         * Render a modal using the specifed path and optionally set a model.
+         * @param modalPath the path to use for the controller and template.
+         * @param model (optional) the model to set on the controller for the modal.
+         */
+        openModal: function(modalPath, model) {
+            if (model) {
+                this.controllerFor(modalPath).set('model', model);
+            }
+            this.renderModal(modalPath);
+        },
+        /**
+         * Update an open modal using the specifed model.
+         * @param modalPath the path to use for the controller and template.
+         * @param model (optional) the model to set on the controller for the modal.
+         */
+        updateModal: function(modalPath, model) {
+            this.controllerFor(modalPath).set('model', model);
         }
     },
 
     model: function() {
-        return this.store.find('config');        
+        return this.store.find('config');
     },
-    
+
     afterModel: function(resolvedModel) {
         this.controllerFor('navigation').set('allowSearch',false);
         if (resolvedModel) {
@@ -24,7 +49,14 @@ var ApplicationRoute = Ember.Route.extend(Ember.SimpleAuth.ApplicationRouteMixin
                 this.use_google_auth = use_google_auth.get('value');
             }
         }
-    }
-    
+    },
+
+    renderModal: function(template) {
+        this.render(template, {
+            into: 'application',
+            outlet: 'modal'
+        });
+    },
+
 });
 export default ApplicationRoute;
