+++ app/templates/sidebar_nav/header.hbs
@@ -13,6 +13,7 @@
         {{else}}
             {{#link-to 'login'}}Login{{/link-to}}
         {{/if}}
+          <a href="#" {{ action 'about' }}>About HospitalRun</a>
         </nav>
     {{/if}}
